Change timestamp double to int (64 bit) type

+++ src/publish_fake.py
@@ -224,7 +224,7 @@ if __name__ == "__main__":
                 )
 
                 row = {
-                    "timestamp": epoch.timestamp() * 1e6,
+                    "timestamp": int(epoch.timestamp() * 1e6),
                     **key_vals,
                     **fake.values(),
                 }
